Move config to separate file and make NullConfig inherit from Config
